only asm customer list
